% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BentCableVarEst.R
\name{CableProduce}
\alias{CableProduce}
\title{BentCable method of parameters stand error estimation for multi-kink quantile regression}
\usage{
CableProduce(x, psi, h)
}
\arguments{
\item{x}{A numeric variable with change point}

\item{psi}{Change points estimation of multi-kink quantile regression}

\item{h}{bandwidth for kernel function}

\item{y}{A vector of response variable}

\item{z}{A vector of covariates (without change point)}

\item{tau}{quantile level}

\item{bet}{Coefficients estimation results of multi-kink quantile regression}
}
\value{
A list with the elements
\item{bet.se}{The estimated standard error of the regression coefficients.}
\item{psi.se}{The estimated standard error of the change points.}
}
\description{
This function gives BentCable method of parameters stand error estimation for multi-kink
quantile regression.
}
\author{
Yu Yue
}
\keyword{BentCable}
\keyword{error}
\keyword{stand}
