% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSQregIter.R
\name{S.func}
\alias{S.func}
\title{BSQreg method of parameters estimation for multi-kink quantile regression}
\usage{
S.func(x, alpha, tau)
}
\arguments{
\item{x}{A numeric variable with change point}

\item{alpha}{parameter of new loss function}

\item{tau}{quantile level}

\item{y}{A vector of response variable}

\item{z}{A vector of covariates (without change point)}

\item{h}{bandwidth for kernel function}

\item{bet.ini}{A initial vector of regression coefficients}

\item{bp.ini}{A initial value vector of change points}

\item{tol}{iterative tolerance value, 1e-8 for default}

\item{max.iter}{the maximum iteration steps, 5000 for default}

\item{yta}{learning rate for iteration,0.1 for default}

\item{a}{learning rate adjustment parameters (reduce learning rate),0.5 for default}

\item{b}{learning rate adjustment parameters (increase learning rate),2 for default}
}
\value{
A list with the elements
\item{psi}{The estimated change points.}
\item{coefficients}{The estimated regression coefficients with intercept.}
\item{iter}{The iteration steps of algorithm.}
\item{Loss.result}{End value of loss function.}
}
\description{
This function gives BSQreg method of parameters estimation for multi-kink
quantile regression.
}
\author{
Yu Yue
}
\keyword{BSQreg}
\keyword{Boosting}
