% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSQregVarEst.R
\name{BSQVar}
\alias{BSQVar}
\title{BSQreg method of parameters stand error estimation for multi-kink quantile regression}
\usage{
BSQVar(y, x, z, tau, h, bet, psi)
}
\arguments{
\item{y}{A vector of response variable}

\item{x}{A numeric variable with change point}

\item{z}{A vector of covariates (without change point)}

\item{tau}{quantile level}

\item{h}{bandwidth for kernel function}

\item{bet}{Coefficients estimation results of multi-kink quantile regression}

\item{psi}{Change points estimation of multi-kink quantile regression}
}
\value{
A list with the elements
\item{bet.se}{The estimated standard error of the regression coefficients.}
\item{psi.se}{The estimated standard error of the change points.}
}
\description{
This function gives BSQreg method of parameters stand error estimation for multi-kink
quantile regression.
}
\author{
Yu Yue
}
\keyword{BSQreg}
\keyword{error}
\keyword{stand}
