% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MKsimdat.R
\name{MKsimdat}
\alias{MKsimdat}
\title{Simulated data from the multi-kink quantile regression}
\usage{
MKsimdat(
  n,
  bet0,
  psi0,
  tau,
  etype = c("normal", "student", "laplace"),
  scedast = c("homo", "hetero")
)
}
\arguments{
\item{n}{sample size.}

\item{bet0}{the vector of real multi-kink regression coefficients.}

\item{psi0}{the real location vector of the change points.}

\item{tau}{the quantile level.}

\item{etype}{type of error, where
etype = "normal" for N(0,1)
etype = "student" for t(3)
etype = "laplace" for LA(0,1)}

\item{scedast}{type of scedasticity, where
scedast = "homo" for homoscedasticity
scedast = "hetero" for heteroscedasticity}
}
\value{
A list with the elements
\item{y}{The response variable.}
\item{x}{The scalar covariate with change points.}
\item{z}{A vector of covariates.}
}
\description{
The function generates data from the multi-kink quantile regression model.
}
\details{
The bent line quantile regression model:
Q_tau (Y|x,z) = [1, x, z,(x-\delta_1)_+,\cdots,(x-\delta_K)_+] * bet0.
}
\author{
Yu Yue
}
\keyword{MKsimdat}
